requirements.txt
fastapi==0.112.*
uvicorn[standard]==0.30.*
asyncpg==0.29.*
python-dotenv==1.0.*
